binder class now handles multiple callbacks for a single binder
changed interface slightly to accomodate this. The Binder init function now has the following signature:
Binder.__init__(self, eventtype, attr_filter, *callbacks)

It is legal to specify no callbacks for the Binder, and the callbacks can be modified at any time through the Binder.funcs attribute.
NOTE: the exception to this is while one of the callbacks is being invoked, since at this time the funcs attribute is being iterated over, and this is not safe