added the kill method for event.Handler objects.
event binders will keep an object alive if you let it go out of scope. The kill method will remove the object from event binders,
so it can safely go out of scope and be deleted. I've tried using weakrefs inside the event binders, but they
die almost immedeately. This is a preferable solution, if it can be made to work