because the render event must have an interpolate event, a clock object must always exist. The clock must be modified for this.
The clock has a set_tps function, because it must be created by the gunge module. This function allows you to update the tps after
gunge creates the clock for you. The clock also has no effect if it is not started, which allows you to have games without the clock object, if you desire